refactor: rename _VersioningCLI to _VersionCLI and move _CoreCLI to src/taipy/core
